# docs/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : true,
            "type" : "object"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "details", "message", "traceId" ],
        "type" : "object"
      },
...
      "HandoffRequest" : {
        "properties" : {
          "targetClientId" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "targetClientId" ],
        "type" : "object"
      },
...
      "JwkSet" : {
        "properties" : {
          "keys" : {
            "items" : {
              "$ref" : "#/components/schemas/PublicJwk"
            },
            "type" : "array"
          }
        },
        "required" : [ "keys" ],
        "type" : "object"
      },
...
      "OpenIdConfiguration" : {
        "properties" : {
          "authorization_endpoint" : {
            "type" : "string"
          },
          "code_challenge_methods_supported" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "end_session_endpoint" : {
            "type" : "string"
          },
          "grant_types_supported" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "id_token_signing_alg_values_supported" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "issuer" : {
            "type" : "string"
          },
          "jwks_uri" : {
            "type" : "string"
          },
          "response_types_supported" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "revocation_endpoint" : {
            "type" : "string"
          },
          "scopes_supported" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "subject_types_supported" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "token_endpoint" : {
            "type" : "string"
          },
          "token_endpoint_auth_methods_supported" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "userinfo_endpoint" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "AgilityHub Core API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/.well-known/jwks.json" : {
      "get" : {
        "description" : "ANON, global. RSA public keys only; no private key material.",
        "operationId" : "wellKnownJwks",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JwkSet"
                }
              }
            },
            "description" : "Public JWK set"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "security" : [ ],
        "summary" : "Get public JWT verification keys",
        "tags" : [ "identity" ]
      }
    },
    "/.well-known/openid-configuration" : {
      "get" : {
        "description" : "ANON. R-01-11; implemented in E1-T05.",
        "operationId" : "discovery",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OpenIdConfiguration"
                }
              }
            },
            "description" : "OIDC discovery metadata"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "security" : [ ],
        "summary" : "Discover the global OpenID provider",
        "tags" : [ "identity" ]
      }
    },
...
    "/api/v1/auth/handoff" : {
      "post" : {
        "description" : "Account token with club context. R-01-13: code lasts 60 seconds and is bound to targetClientId.",
        "operationId" : "handoff",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/HandoffRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HandoffResponse"
                }
              }
            },
            "description" : "Code and destination login URL"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "NO_MEMBERSHIP for the destination role"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Create a one-use code for switching apps",
        "tags" : [ "identity" ]
      }
    },
...
    "/api/v1/me/password" : {
      "put" : {
        "description" : "Account token, never impersonated. R-01-05: current is required only when a password already exists.",
        "operationId" : "password",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Password changed"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "PASSWORD_TOO_SHORT, PASSWORD_MISMATCH, PASSWORD_COMPROMISED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "INVALID_CREDENTIALS for current password"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "FORBIDDEN for impersonated accounts"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Set or change the current account password",
        "tags" : [ "identity" ]
      }
    },
    "/api/v1/me/profile" : {
      "put" : {
        "description" : "Account token with club context. R-01-07. Returns a fresh access token.",
        "operationId" : "profile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description" : "New access token"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "PROFILE_NOT_AVAILABLE"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Select an available club profile",
        "tags" : [ "identity" ]
      }
    },
...
    "/api/v1/members/{id}/impersonation-token" : {
      "post" : {
        "description" : "ADMIN of the same club, never an impersonated token. R-01-09. Target lookup and grant issuance are E1-T04.",
        "operationId" : "impersonation",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ImpersonationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImpersonationTokenResponse"
                }
              }
            },
            "description" : "Impersonation JWT and UTC expiry"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "IMPERSONATION_DENIED"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "NOT_FOUND within the current club"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Create a non-refreshable member impersonation token",
        "tags" : [ "identity" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "https://core.agilitydoghub.com"
  }, {
    "url" : "http://localhost:8080"
  } ]
}